requirements.txt
asgiref==3.2.10
boto3==1.14.47
botocore==1.17.47
Django==3.1
django-classy-tags==1.0.0
django-storages==1.9.1
django-taggit==1.3.0
django-taggit-templatetags2==1.6.1
django-widget-tweaks==1.4.8
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
mysqlclient==2.0.1
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.10
